t07表修改

+++ srv/src/main/java/customer/bean/pch/Pch08Person.java
@@ -1,11 +1,13 @@
 package customer.bean.pch;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class Pch08Person {
     private String id;
-    private Integer personNo;
+    private BigDecimal personNo;
     private String quoNumber;
     private Integer quoItem;
 }
